python3 compatibility ez_setup -> setuptools travis tests from python 2.7 to  3.6
reverted LF <-> CLRF conversion and removed .gitattributes
