+++ .travis.yml
@@ -1,7 +1,14 @@
 language: python
 python:
   - "2.7"
-# command to install dependencies
-install: "python setup.py install"
+  - "3.3"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+  - "pypy"
 # command to run tests
-script: make test
+install:
+  - pip install -U pip setuptools
+  - pip install -U -r test_require.txt
+script:
+  - python setup.py test
