+++ AppiumLibrary/keywords/_screenshot.py
