+++ AppiumLibrary/keywords/_touch.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from appium.webdriver.common.touch_action import TouchAction
+
 from AppiumLibrary.locators import ElementFinder
 from .keywordgroup import KeywordGroup
