+++ AppiumLibrary/utils/__init__.py
@@ -1,5 +1,9 @@
+import six
+
 from .applicationcache import ApplicationCache
 
+if six.PY3:
+    unicode = str
 
 def escape_xpath_value(value):
     value = unicode(value)
