+++ AppiumLibrary/utils/applicationcache.py
