+++ pytest.ini
@@ -1,4 +1,5 @@
-[pytest] 
-python_files = test_*.py
+[pytest]
+testpaths = tests
 norecursedirs = VENV demo
-python_paths = src
+python_paths = AppiumLibrary
+addopts = --cov=AppiumLibrary
