+++ setup.cfg
@@ -1,2 +1,8 @@
 [metadata]
 description-file = README.rst
+
+[bdist_wheel]
+universal=1
+
+[aliases]
+test=pytest
