+++ test_require.txt
@@ -1,0 +1,6 @@
+mock >= 2.0.0
+pytest-cov >= 2.5.1
+pytest-xdist >= 1.16.0
+pytest-pythonpath >= 0.7.1
+pytest >= 3.1.0
+six >= 1.10.0
